+++ components/smart_display/runtime_tiles.h
@@ -1,5 +1,15 @@
 #pragma once
 #include "runtime_model.h"
+#ifdef ESP_SCREEN_HOST
+#include "host_shims.h"
+#else
+#include "esphome/core/preferences.h"
+#include "esphome/components/json/json_util.h"
+#include "esphome/components/api/api_server.h"
+#include "esphome/core/hal.h"
+#include "esphome/core/util.h"
+#include "esphome/core/time.h"
+#endif
 #include "header_bar.h"
 #include "tile_palette.h"
 #include "overlay_card.h"
@@ -8,7 +18,6 @@
 #include "tile_icon.h"
 #include "screen_settings.h"
 #include "settings_screen.h"
-#include "esphome/core/preferences.h"
 #include "climate_card.h"
 #include "screen_input.h"
 #include "light_controls.h"
@@ -20,11 +29,6 @@
 #include "light_card.h"
 #include "weather_card.h"
 #include "swipe_profile.h"
-#include "esphome/components/json/json_util.h"
-#include "esphome/components/api/api_server.h"
-#include "esphome/core/hal.h"
-#include "esphome/core/util.h"
-#include "esphome/core/time.h"
 #include "lvgl.h"
 #include <functional>
 #include <algorithm>